+++ spec/dot_env_validator_spec.rb
@@ -187,6 +187,106 @@ RSpec.describe DotenvValidator do
           end
         end
       end
+
+      context "and there is a uuid format parameter in the comment" do
+        let(:sample_lines) { StringIO.new("KEY_ID=9c382b2a-c84f-4dfd-8bc6-89dc413850bb # format=uuid") }
+
+        context "and ENV variable is a uuid in the default format" do
+          it "returns true" do
+            ClimateControl.modify KEY_ID: "ade67c20-bf67-4b10-ae9a-5809d9e1d041" do
+              expect(DotenvValidator.check).to be_truthy
+            end
+          end
+        end
+
+        context "and the ENV variable is a uuid in the compact format" do
+          it "returns true" do
+            ClimateControl.modify KEY_ID: "ade67c20bf674b10ae9a5809d9e1d041" do
+              expect(DotenvValidator.check).to be_truthy
+            end
+          end
+        end
+
+        context "and the ENV variable has invalid characters" do
+          around do |example|
+            ClimateControl.modify KEY_ID: "xde67c20bf674b10ae9a5809d9e1d041" do
+              example.run
+            end
+          end
+
+          it "returns false" do
+            expect(DotenvValidator.check).to be_falsey
+          end
+
+          it "displays a warning message in STDOUT" do
+            msg = "WARNING - Environment variables with invalid format: KEY_ID\n"
+
+            expect do
+              DotenvValidator.check
+            end.to output(msg).to_stdout
+          end
+        end
+
+        context "and the ENV variable is in an invalid format" do
+          around do |example|
+            ClimateControl.modify KEY_ID: "ade67c20bf674b10a-e9a5809d9e1d041" do
+              example.run
+            end
+          end
+
+          it "returns false" do
+            expect(DotenvValidator.check).to be_falsey
+          end
+
+          it "displays a warning message in STDOUT" do
+            msg = "WARNING - Environment variables with invalid format: KEY_ID\n"
+
+            expect do
+              DotenvValidator.check
+            end.to output(msg).to_stdout
+          end
+        end
+
+        context "and the ENV variable is too short" do
+          around do |example|
+            ClimateControl.modify KEY_ID: "ade67c20bf674b10a809d9e1d041" do
+              example.run
+            end
+          end
+
+          it "returns false" do
+            expect(DotenvValidator.check).to be_falsey
+          end
+
+          it "displays a warning message in STDOUT" do
+            msg = "WARNING - Environment variables with invalid format: KEY_ID\n"
+
+            expect do
+              DotenvValidator.check
+            end.to output(msg).to_stdout
+          end
+        end
+
+        context "and the ENV variable is too long" do
+          around do |example|
+            ClimateControl.modify KEY_ID: "ade67c20bf674b10ae9a5809d9e1d041a89b3d4a567" do
+              example.run
+            end
+          end
+
+          it "returns false" do
+            expect(DotenvValidator.check).to be_falsey
+          end
+
+          it "displays a warning message in STDOUT" do
+            msg = "WARNING - Environment variables with invalid format: KEY_ID\n"
+
+            expect do
+              DotenvValidator.check
+            end.to output(msg).to_stdout
+          end
+        end
+      end
     end
   end
 
@@ -283,6 +383,78 @@ RSpec.describe DotenvValidator do
             msg = "Environment variables with invalid format: MAYBE"
 
             ClimateControl.modify MAYBE: "possibly" do
+              expect do
+                DotenvValidator.check!
+              end.to raise_error(RuntimeError, msg)
+            end
+          end
+        end
+      end
+
+      context "and there is a uuid format parameter in the comment" do
+        let(:sample_lines) { StringIO.new("KEY_ID=9c382b2a-c84f-4dfd-8bc6-89dc413850bb # format=uuid") }
+
+        context "and ENV variable is a uuid in the default format" do
+          it "does not raise a runtime error" do
+            ClimateControl.modify KEY_ID: "ade67c20-bf67-4b10-ae9a-5809d9e1d041" do
+              expect do
+                DotenvValidator.check!
+              end.not_to raise_error
+            end
+          end
+        end
+
+        context "and the ENV variable is a uuid in the compact format" do
+          it "does not raise a runtime error" do
+            ClimateControl.modify KEY_ID: "ade67c20bf674b10ae9a5809d9e1d041" do
+              expect do
+                DotenvValidator.check!
+              end.not_to raise_error
+            end
+          end
+        end
+
+        context "and the ENV variable has invalid characters" do
+          it "raises a runtime error with a warning message" do
+            msg = "Environment variables with invalid format: KEY_ID"
+
+            ClimateControl.modify KEY_ID: "xde67c20bf674b10ae9a5809d9e1d041" do
+              expect do
+                DotenvValidator.check!
+              end.to raise_error(RuntimeError, msg)
+            end
+          end
+        end
+
+        context "and the ENV variable is in an invalid format" do
+          it "raises a runtime error with a warning message" do
+            msg = "Environment variables with invalid format: KEY_ID"
+
+            ClimateControl.modify KEY_ID: "ade67c20bf674b10ae-9a5809d9e1d041" do
+              expect do
+                DotenvValidator.check!
+              end.to raise_error(RuntimeError, msg)
+            end
+          end
+        end
+
+        context "and the ENV variable is too short" do
+          it "raises a runtime error with a warning message" do
+            msg = "Environment variables with invalid format: KEY_ID"
+
+            ClimateControl.modify KEY_ID: "ade67c20bf674b10809d9e1d041" do
+              expect do
+                DotenvValidator.check!
+              end.to raise_error(RuntimeError, msg)
+            end
+          end
+        end
+
+        context "and the ENV variable is too long" do
+          it "raises a runtime error with a warning message" do
+            msg = "Environment variables with invalid format: KEY_ID"
+
+            ClimateControl.modify KEY_ID: "ade67c20bf674b10ae9a5809d9e1d041a89b3d4a567" do
               expect do
                 DotenvValidator.check!
               end.to raise_error(RuntimeError, msg)